Email authentication , responsive layout

+++ server/Controller/SubCategory.controller.js
@@ -15,7 +15,9 @@ exports.addSubCategory = asyncHandler(async (req, res) => {
     const createRecords = new SubCategoryModel(data);
     const result = await createRecords.save();
     if (result) {
-      res.status(201).json({ message: "data has been added successfully !" });
+      res
+        .status(201)
+        .json({ message: "SubCategory data added successfully !" });
     }
   } catch (err) {
     console.error("Sub category data not saved", err);
@@ -27,33 +29,53 @@ exports.addSubCategory = asyncHandler(async (req, res) => {
 exports.getSubCategoryData = asyncHandler(async (req, res) => {
   try {
     const data = await SubCategoryModel.find();
-    if (data.length > 0) {
-      const transformdata = data.map((data, index) => ({
-        "sl.no": data._id,
-        "Department Name": data.departmentName,
-        "Category Name": data.categoryName,
-        "SubCategory Name": data.subCategoryName,
-        "Short Name": data.shortName,
-        "Is Active": data.isActive,
-        "image Url": data.imageUrl,
-      }));
-      res
-        .status(200)
-        .json({ message: "data fetched successfully !", data: transformdata });
+    if (!data) {
+      return res
+        .status(204)
+        .json({ message: "sub Category data not availabel !" });
     }
-  } catch (err) {
-    console.error("Failed retrieving sub category data", err);
+    const transformdata = data?.map((data) => ({
+      "sl.no": data._id,
+      "Department Name": data.departmentName,
+      "Category Name": data.categoryName,
+      "SubCategory Name": data.subCategoryName,
+      "Short Name": data.shortName,
+      "Is Active": data.isActive,
+      "image Url": data.imageUrl,
+    }));
+    res.status(200).json({
+      message: "Sub-Category data fetched successfully !",
+      data: transformdata,
+    });
+  } catch (error) {
+    console.error("Failed retrieving sub category data", error);
+    return res.status(500).json({
+      message: "An internal server error occurred. Please try again later",
+    });
   }
 });
 exports.retrieveIndividualRecord = asyncHandler(async (req, res) => {
   try {
     const { id } = req.params;
-    const record = await SubCategoryModel.findById(id);
-    if (record !== null) {
-      return res.status(200).json(record);
+    const data = await SubCategoryModel.findById(id);
+    if (!data) {
+      return res.status(400).json({ message: "data not found." });
     }
+    const transformdata = {
+      "sl.no": data._id,
+      "Department Name": data.departmentName,
+      "Category Name": data.categoryName,
+      "SubCategory Name": data.subCategoryName,
+      "Short Name": data.shortName,
+      "Is Active": data.isActive,
+      "image Url": data.imageUrl,
+    };
+    return res.status(200).json({
+      message: "subCategory data has been retrieved!",
+      data: transformdata,
+    });
   } catch (err) {
-    console.error("Sub category data not saved", err);
+    console.error("failed to get Sub category data ", err);
     res
       .status(500)
       .json({ message: "Unknown error occured at server side", err });
@@ -77,11 +99,15 @@ exports.currentAvailabelDepartments = asyncHandler(async (req, res) => {
       },
     ];
     const uniqueDepartments = await CategoryModel.aggregate(stages);
-    if (uniqueDepartments.length > 0) {
-      res
-        .status(200)
-        .json({ message: "Availabel departments", data: uniqueDepartments });
+    if (!uniqueDepartments) {
+      return res
+        .status(204)
+        .json({ message: "Departments currently not availabel" });
     }
+    return res.status(200).json({
+      message: "currently availabel departments",
+      data: uniqueDepartments,
+    });
   } catch (err) {
     res
       .status(500)
@@ -95,7 +121,7 @@ exports.getCategoriesByDepartment = async (req, res) => {
     const categories = (
       await CategoryModel.find({ departmentName: department })
     ).map((data) => data.categoryName);
-    if (!categories || categories.length === 0) {
+    if (!categories) {
       return res
         .status(404)
         .json({ message: "No categories found for this department" });
@@ -110,41 +136,52 @@ exports.getCategoriesByDepartment = async (req, res) => {
 exports.deleteRecord = asyncHandler(async (req, res) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid ID format" });
+    }
     const result = await SubCategoryModel.findByIdAndDelete(id, {
       new: true,
     });
-    if (result !== null) {
-      return res.status(200).json("Sub Category deleted !");
-    } else {
-      res.status(400).json({ message: `Sub Category not deleted !` });
+    if (!result) {
+      return res.status(400).json({ message: `Sub Category not deleted !` });
     }
+    return res.status(200).json({ message: "record has been deleted !" });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Server error", error: error.message });
   }
 });
 exports.updateSubCategory = asyncHandler(async (req, res) => {
-  const rowid = req.params.id;
-  if (!mongoose.Types.ObjectId.isValid(rowid)) {
-    return res.status(400).json({ message: "Invalid ID format" });
-  }
-  const body = {
-    departmentName: req.body["Department Name"],
-    categoryName: req.body["Category Name"],
-    subCategoryName: req.body["SubCategory Name"],
-    shortName: req.body["Short Name"],
-    isActive: req.body["Is Active"],
-    imageUrl: req.body["image Url"],
-  };
-  const updateDocument = await SubCategoryModel.findByIdAndUpdate(
-    rowid,
-    { $set: body },
-    { new: true, runValidators: true }
-  );
-  if (updateDocument) {
-    return res.status(200).json({
-      message: "Sub category has been updated!",
-      data: updateDocument,
+  try {
+    console.log("body", req.body);
+    const rowid = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(rowid)) {
+      return res.status(400).json({ message: "Invalid ID format" });
+    }
+    const body = {
+      departmentName: req.body["Department Name"],
+      categoryName: req.body["Category Name"],
+      subCategoryName: req.body["SubCategory Name"],
+      shortName: req.body["Short Name"],
+      isActive: req.body["Is Active"],
+      imageUrl: req.body["image Url"],
+    };
+    const updateDocument = await SubCategoryModel.findByIdAndUpdate(
+      rowid,
+      { $set: body },
+      { new: true, runValidators: true }
+    );
+    if (updateDocument) {
+      return res.status(200).json({
+        message: "Sub category has been updated!",
+        data: updateDocument,
+      });
+    }
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({
+      message: "Internal error has been occurred",
+      error: error.message,
     });
   }
 });
